Add unified CI workflow covering all platforms and Go Guardian
Replace the five per-platform workflow files with a single ci.yml that builds
Linux (native + cross matrix), Windows, macOS, Android, and Go Guardian in one
pipeline, plus related build-system fixes discovered while stabilising it.

CI workflow:
- Add unified .github/workflows/ci.yml; disable legacy per-platform workflows.
- Cache third-party deps (boost, openssl, jemalloc) outside the source tree and
  expose them via cmake -D flags instead of sed-rewriting CMakeLists.txt.
- Linux cross matrix: correct apt package names, sanitize platform names with
  slashes when composing artifact filenames.
- Android: export PPP_ANDROID_ABI (required by android/CMakeLists.txt); fix sed
  quoting; correct elif syntax; guard incompatible APIs; add platform macros.
- Windows: manual vcpkg install of required boost components; add setup-node.

Build system:
- CMakeLists.txt: remove bare main.cpp from FILE(GLOB_RECURSE ...) so vendored
  third-party/**/main.cpp (e.g. Boost.Asio C++20 examples) are not pulled into
  the ppp target; append it explicitly instead.
- build-openppp2-by-cross.sh: drive configuration via cmake -D instead of
  mutating CMakeLists.txt with sed.

Application:
- ppp/app/ApplicationInitialize.cpp: split the non-Windows branch into _LINUX
  vs generic Unix so the VirtualEthernetSwitcher ctor is called with the right
  overload on macOS/Android (no SsmtMQ).

Go Guardian:
- Split procattr into _windows / _other files and add kill_{unix,windows}.go
  so the guardian builds cleanly on Windows as well as Unix.
