+++ go/guardian/instance/procattr_windows.go
@@ -1,9 +1,9 @@
-//go:build !linux && !windows
+//go:build windows
 
 package instance
 
 import "syscall"
 
 func procAttrForPTY(_ uintptr) *syscall.SysProcAttr {
-	return &syscall.SysProcAttr{Setpgid: true}
+	return &syscall.SysProcAttr{}
 }
